Workaround for problem where spurious PulseBlaster falling edges caused the novatech to advance its table, despite not being in table mode.
The workaround has it transition to table mode and back again, before going into
table mode for a buffered run. This "double clutch" mechanism appears to reset
its memory pointer to the start of the table. There may still be a potential race
condition in which the PulseBlaster gives its pulses after the double-clutch.
If so, A fix for this might entail imposing an ordering on the devices transitioning
to buffered.